processamento_imagens/pyproject.toml
[tool.poetry]
name = "processamento-imagens"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
numpy = "^2.2.0"
pyyaml = "^6.0.2"
matplotlib = "^3.9.3"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
